refactor: :recycle: refactor sources and destinations modules
