+++ ui/package.json
@@ -14,7 +14,6 @@
 	},
 	"dependencies": {
 		"@phosphor-icons/react": "^2.1.7",
-		"@rjsf/core": "^5.24.8",
 		"@rjsf/utils": "^5.24.8",
 		"antd": "^5.24.2",
 		"axios": "^1.8.1",
@@ -27,7 +26,6 @@
 	},
 	"devDependencies": {
 		"@eslint/js": "^9.19.0",
-		"@rjsf/validator-ajv8": "^5.24.8",
 		"@types/json-schema": "^7.0.15",
 		"@types/node": "^22.13.10",
 		"@types/react": "^18.2.62",
