+++ ui/src/modules/common/components/SetupTypeSelector.tsx
@@ -1,0 +1,27 @@
+import { Radio } from "antd"
+import { SetupType, SetupTypeSelectorProps } from "../../../types"
+
+export const SetupTypeSelector: React.FC<SetupTypeSelectorProps> = ({
+	value,
+	onChange,
+	newLabel = "Set up a new source",
+	existingLabel = "Use an existing source",
+}) => {
+	return (
+		<div className="mb-4 flex">
+			<Radio.Group
+				value={value}
+				onChange={e => onChange(e.target.value as SetupType)}
+				className="flex"
+			>
+				<Radio
+					value="new"
+					className="mr-8"
+				>
+					{newLabel}
+				</Radio>
+				<Radio value="existing">{existingLabel}</Radio>
+			</Radio.Group>
+		</div>
+	)
+}
