+++ ui/src/modules/destinations/pages/DestinationEdit.tsx
@@ -9,33 +9,19 @@ import { destinationService } from "../../../api/services/destinationService"
 import { jobService } from "../../../api"
 import StepTitle from "../../common/components/StepTitle"
 import DeleteModal from "../../common/Modals/DeleteModal"
-import AWSS3 from "../../../assets/AWSS3.svg"
-import ApacheIceBerg from "../../../assets/ApacheIceBerg.svg"
 import {
 	getCatalogInLowerCase,
 	getConnectorImage,
 	getConnectorName,
 } from "../../../utils/utils"
 import EditDestinationModal from "../../common/Modals/EditDestinationModal"
-import { Entity, EntityJob } from "../../../types"
+import { DestinationJob, Entity } from "../../../types"
 import type { ColumnsType } from "antd/es/table"
 import { formatDistanceToNow } from "date-fns"
 import TestConnectionSuccessModal from "../../common/Modals/TestConnectionSuccessModal"
 import TestConnectionFailureModal from "../../common/Modals/TestConnectionFailureModal"
 import TestConnectionModal from "../../common/Modals/TestConnectionModal"
-
-// Define the job structure for destination jobs table
-interface DestinationJob extends Omit<EntityJob, "last_runtime"> {
-	source_type: string
-	last_run_time: string
-	last_run_state: string
-	source_name: string
-}
-
-interface ConnectorOption {
-	value: string
-	label: React.ReactNode
-}
+import { connectorOptions } from "../components/connectorOptions"
 
 interface CatalogOption {
 	value: string
@@ -99,36 +85,6 @@ const DestinationEdit: React.FC<DestinationEditProps> = ({
 	} = useAppStore()
 
 	const navigate = useNavigate()
-
-	// Define connector options
-	const connectorOptions: ConnectorOption[] = [
-		{
-			value: "Amazon S3",
-			label: (
-				<div className="flex items-center">
-					<img
-						src={AWSS3}
-						alt="AWS S3"
-						className="mr-2 size-5"
-					/>
-					<span>Amazon S3</span>
-				</div>
-			),
-		},
-		{
-			value: "Apache Iceberg",
-			label: (
-				<div className="flex items-center">
-					<img
-						src={ApacheIceBerg}
-						alt="Apache Iceberg"
-						className="mr-2 size-5"
-					/>
-					<span>Apache Iceberg</span>
-				</div>
-			),
-		},
-	]
 
 	// Define catalog options
 	const catalogOptions: CatalogOption[] = [
