+++ ui/src/modules/sources/pages/SourceEdit.tsx
@@ -13,25 +13,11 @@ import EditSourceModal from "../../common/Modals/EditSourceModal"
 import { sourceService } from "../../../api"
 import { formatDistanceToNow } from "date-fns"
 import { jobService } from "../../../api"
-import { Entity } from "../../../types"
+import { Entity, SourceJob } from "../../../types"
 import TestConnectionSuccessModal from "../../common/Modals/TestConnectionSuccessModal"
 import TestConnectionFailureModal from "../../common/Modals/TestConnectionFailureModal"
 import TestConnectionModal from "../../common/Modals/TestConnectionModal"
-
-interface SourceJob {
-	destination_type: string
-	last_run_time: string
-	last_run_state: string
-	id: number
-	name: string
-	activate: boolean
-	destination_name: string
-}
-
-interface ConnectorOption {
-	value: string
-	label: React.ReactNode
-}
+import connectorOptions from "../components/connectorOptions"
 
 interface SourceEditProps {
 	fromJobFlow?: boolean
@@ -69,9 +55,7 @@ const SourceEdit: React.FC<SourceEditProps> = ({
 
 	const {
 		sources,
-		// jobs,
 		fetchSources,
-		// fetchJobs,
 		updateSource,
 		setShowEditSourceModal,
 		setShowTestingModal,
@@ -79,49 +63,6 @@ const SourceEdit: React.FC<SourceEditProps> = ({
 		setShowFailureModal,
 		setSourceTestConnectionError,
 	} = useAppStore()
-
-	// Create connector options once
-	const connectorOptions: ConnectorOption[] = [
-		{
-			value: "MongoDB",
-			label: (
-				<div className="flex items-center">
-					<img
-						src={getConnectorImage("MongoDB")}
-						alt="MongoDB"
-						className="mr-2 size-5"
-					/>
-					<span>MongoDB</span>
-				</div>
-			),
-		},
-		{
-			value: "Postgres",
-			label: (
-				<div className="flex items-center">
-					<img
-						src={getConnectorImage("Postgres")}
-						alt="Postgres"
-						className="mr-2 size-5"
-					/>
-					<span>Postgres</span>
-				</div>
-			),
-		},
-		{
-			value: "MySQL",
-			label: (
-				<div className="flex items-center">
-					<img
-						src={getConnectorImage("MySQL")}
-						alt="MySQL"
-						className="mr-2 size-5"
-					/>
-					<span>MySQL</span>
-				</div>
-			),
-		},
-	]
 
 	useEffect(() => {
 		fetchSources()
@@ -290,7 +231,6 @@ const SourceEdit: React.FC<SourceEditProps> = ({
 	const handleDelete = () => {
 		if (!source) return
 
-		// Create a simplified source object for deletion
 		const sourceToDelete = {
 			...source,
 			name: sourceName || source.name,
