+++ ui/src/modules/sources/pages/Sources.tsx
@@ -3,10 +3,10 @@ import { Button, Tabs, Empty, message, Spin } from "antd"
 import { useNavigate } from "react-router-dom"
 import { useAppStore } from "../../../store"
 import SourceTable from "../components/SourceTable"
-import FirstSource from "../../../assets/FirstSource.svg"
-import SourcesTutorial from "../../../assets/SourcesTutorial.svg"
-import { LinktreeLogo, PlayCircle, Plus } from "@phosphor-icons/react"
+import { LinktreeLogo, Plus } from "@phosphor-icons/react"
 import { Entity } from "../../../types"
+import { sourceTabs } from "../../../utils/constants"
+import SourceEmptyState from "../components/SourceEmptyState"
 
 const Sources: React.FC = () => {
 	const [activeTab, setActiveTab] = useState("active")
@@ -66,11 +66,6 @@ const Sources: React.FC = () => {
 
 	const showEmpty = !isLoadingSources && sources.length === 0
 
-	const sourceTabs = [
-		{ key: "active", label: "Active sources" },
-		{ key: "inactive", label: "Inactive sources" },
-	]
-
 	if (sourcesError) {
 		return (
 			<div className="p-6">
@@ -122,50 +117,11 @@ const Sources: React.FC = () => {
 							/>
 						</div>
 					) : tab.key === "active" && showEmpty ? (
-						<div className="flex flex-col items-center justify-center py-16">
-							<img
-								src={FirstSource}
-								alt="Empty state"
-								className="mb-8 h-64 w-96"
-							/>
-							<div className="mb-2 text-blue-600">Welcome User !</div>
-							<h2 className="mb-2 text-2xl font-bold">
-								Ready to create your first source
-							</h2>
-							<p className="mb-8 text-gray-600">
-								Get started and experience the speed of OLake by running jobs
-							</p>
-							<Button
-								type="primary"
-								className="border-1 mb-12 border-[1px] border-[#D9D9D9] bg-white px-6 py-4 text-black"
-								onClick={handleCreateSource}
-							>
-								<Plus />
-								New Source
-							</Button>
-							<div className="w-[412px] rounded-xl border-[1px] border-[#D9D9D9] bg-white p-4 shadow-sm">
-								<div className="flex items-center gap-4">
-									<img
-										src={SourcesTutorial}
-										alt="Job Tutorial"
-										className="rounded-lg"
-									/>
-									<div className="flex-1">
-										<div className="mb-1 flex items-center gap-1 text-xs">
-											<PlayCircle color="#9f9f9f" />
-											<span className="text-[#9F9F9F]">OLake/ Tutorial</span>
-										</div>
-										<div className="text-xs">
-											Checkout this tutorial, to know more about running jobs
-										</div>
-									</div>
-								</div>
-							</div>
-						</div>
+						<SourceEmptyState handleCreateSource={handleCreateSource} />
 					) : filteredSources().length === 0 ? (
 						<Empty
 							image={Empty.PRESENTED_IMAGE_SIMPLE}
-							description="No data"
+							description="No sources configured"
 							className="flex flex-col items-start"
 						/>
 					) : (
