+++ ui/src/types/formTypes.ts
@@ -1,0 +1,6 @@
+export interface FormFieldProps {
+	label: string
+	required?: boolean
+	children: React.ReactNode
+	error?: string | null
+}
