+++ ui/src/types/index.ts
@@ -4,3 +4,6 @@ export * from "./entityTypes"
 export * from "./jobTypes"
 export * from "./streamTypes"
 export * from "./apiTypes"
+export * from "./sourceTypes"
+export * from "./destinationTypes"
+export * from "./formTypes"
