+++ ui/src/types/sourceTypes.ts
@@ -1,0 +1,41 @@
+import { Entity, EntityBase } from "./entityTypes"
+
+export interface SourceTableProps {
+	sources: Entity[]
+	loading: boolean
+	onEdit: (id: string) => void
+	onDelete: (source: Entity) => void
+}
+
+export interface Source {
+	id: string | number
+	name: string
+	type: string
+	version: string
+	config?: any
+}
+
+export interface CreateSourceProps {
+	fromJobFlow?: boolean
+	onComplete?: () => void
+	stepNumber?: string
+	stepTitle?: string
+	initialConfig?: EntityBase
+	initialFormData?: any
+	initialName?: string
+	initialConnector?: string
+	onSourceNameChange?: (name: string) => void
+	onConnectorChange?: (connector: string) => void
+	onFormDataChange?: (formData: any) => void
+	onVersionChange?: (version: string) => void
+}
+
+export interface SourceJob {
+	destination_type: string
+	last_run_time: string
+	last_run_state: string
+	id: number
+	name: string
+	activate: boolean
+	destination_name: string
+}
